Stage 1/6: Elementary!
Description

Today, XML and JSON are the most popular formats of data that can be understood both by computers and by humans. These formats are used to exchange information between servers, between various components in a single application, and much more.

Before we start, let’s explore what XML and JSON are.

XML stands for eXtensible Markup Language.
XML uses tags to hold data. Everything in XML is built on tags.

A tag is a specially-formatted label used in markup languages: <opening_tag>content</closing tag>. The content of a tag is inside the same opening and closing tags. The inside of the tag can be empty. For example <html></html>. In this case, it is possible to shorthand this into one opening and closing tag. For example, <html/> is an empty tag with no corresponding pair.

XML is used for storing and transporting data.
XML contains data in a hierarchical order.

Objectives

In this stage, you will write a simple XML/JSON converter.

Conversion of XML to JSON:

Extract an element between tags and enclose it within double quotes.
Extract element content and enclose it within double quotes. If the element is empty, assign null (JSON value).

Conversion of JSON to XML:

Extract key between double quotes (XML element).
Extract element content from JSON value.
If value is not null, enclose JSON with <></>, else enclose within < />.

Write a program that reads XML/JSON from standard input and outputs the converted XML/JSON.

If the first non-space character of the input is < then this is an XML document, and you need to convert it into JSON. Otherwise, the first non-space character would be { and you need to convert this JSON into an XML. Before checking, the grader deletes all spaces and newline characters so you can output it pretty much as you wish. Try to output this in as readable a format as you can.