Stage 4/6: Parsing JSON
Description
In the previous stage, you've expanded an XML file, and in this stage you will expand a JSON file.

To parse it you should write a recursive function that will check if the JSON value starts with a curly brace that can be presented as a start of child nodes and continue to check subkeys.

JSON with sub-elements would look like this:

{
    "key": {
        "child_key1": "child_key_value",
        "child_key2": "child_key_value",
        "child_key3": {
            "child_child_key": "value"
        },
        "child_key4": {
            "child_child_key1": "value1",
            "child_child_key2": "value2"
        }
    }
}
The algorithm for traversing JSON is similar to traversing XML.

Let’s add some magic constants @ and # to represent attributes and values. Now the JSON tree would look like this:

{
    "key": {
        "child_key1": {
            "@attribute1": "value1",
            "@attribute2": "value2",
            "#child_key1": "value3"
        },
        "child_key2": "child_key_value",
        "child_key3": {
            "@attribute1": "value4",
            "@attribute2": "value5",
            "#child_key3": null
        },
        "child_key4": {
            "child_child_key1": "value1",
            "child_child_key2": "value2"
        }
    }
}
Parsing JSON attributes is different from parsing XML. As you can see, every attribute and corresponding value can be presented as child_nodes for every element. That means you just need to expand your traversal of the JSON tree by checking if its subchildren start with @ or #.

However, there is a problem: what would you do if the value contains keys that start with @ but there are no keys that start with #? Similarly, what would you do if the value contains keys that start with # along with keys that start with @ but there are also keys that start with something else?

In these cases, you should forget about the special meaning of @ and # in front of a key and look at them as regular keys with regular values. You should remove @ and # from the key as these symbols are not allowed in the XML tag names.

Also, if the key contains only the @ symbol, then it's not considered a special key representing an attribute.

For the keys that lost their meaning due to some errors, you should omit @ and # symbols from the start as XML elements can't start with these symbols. If there are @key and key in the object, you should prefer a key without the @ symbol.

Also, avoid empty keys. Do not map them to the output as they definitely would break something during conversion to an XML. Keep in mind that if the key contains only the @ symbol, then it's not considered a special key representing an attribute and after omitting this symbol key would be empty. Therefore, you shouldn't convert keys @ and #.

Also, if the attribute is null, change it to the empty string.

So, in short, you need to consider a JSON element as an object with attributes only if:

1. The object has a key with the same name as the object, with a # symbol in front of it. For example, if the key of the object is "obj" then the value of this object has to be inside "#obj" key inside this object. Note that if such a key does not exist in the object, this object should not be considered a single XML object with attributes.

2. The value object contains the #value key and all other attributes begin with the @ symbol and are longer than 1 character. If this object has at least one key that equals @ or does not start with @ (except #value), then this object should not be considered a single XML object with attributes.

3. If the value of any key starting with @ is not a number, string or null (in other words, it will be an object starting with "{"), then this object cannot be an attribute of a single XML object and the @ symbol should be removed from this key, and thus the object cannot be considered a single XML object.

The object without attributes would look like a standard JSON object, without @ or # symbols around. In particular, all wrong cases of creating an object with attributes result in a standard object after certain manipulations. For example inner1, inner2 and inner3 from the first example are objects without attributes with no errors in construction.

Some cases from the first example:

1. inner4 contains a key @ and it is not an attribute. That's why this object is not considered a single object with attributes. But since @ key is not allowed, this key would be removed and #inner4 key will turn into inner4 key. After all the manipulations, this object will look like this:

"inner4": {
    "inner4": "value3"
}
Same rules apply to inner4.2.

2. inner5 contains key #inner4 but should contain key #inner5 to be a single object with attributes.

3. inner6 constructed in the correct way, with #inner6 key and an attribute that starts with @. That's why inner6 is a valid single XML object with attributes.

4. inner7 constructed in the correct way, with #inner7 key, just without attributes and that's allowed.

5. inner8 doesn't contain key #inner8 and should not be considered a single XML object with attributes.

6. inner11 contains children, all their names are not allowed, so these children will be deleted and inner11 will become an empty object. Similar to empty2, the value of this object has to be an empty string.

All of the variants are described in the output example below.

In this stage you should write a program that reads JSON file from disk, parses it and outputs the following:

Key:
path = super_key … parent_key
value = value
attributes : [ if key doesn’t contain attributes skip this ]
attributeN = attributeN
Looking ahead, how are you going to convert this XML into JSON?

<elem1 attr1="val1" attr2="val2">
    <elem2 attr3="val3" attr4="val4">Value1</elem2>
    <elem3 attr5="val5" attr6="val6">Value2</elem3>
</elem1>
Note that the element elem1 contains attributes, but its value is on another level of complexity, not just a number, null, or a string. Well, in this case you can also use standard rules. See how it should be converted into JSON:

{
    "elem1": {
        "@attr1": "val1",
        "@attr2": "val2",
        "#elem1": {
            "elem2": {
                "@attr3": "val3",
                "@attr4": "val4",
                "#elem2": "Value1"
            },
            "elem3": {
                "@attr5": "val5",
                "@attr6": "val6",
                "#elem3": "Value2"
            }
        }
    }
}
Also, you should not rely on new lines, they are only there to make the code more readable. Take a look at the last output example: the initial JSON file doesn't contain any new lines, everything is on a single line although the content is equal to the last but one example.

You should read from the file named test.txt. If you want to test your program, you should check it on some other file, because the contents of this file will be overwritten during testing and after testing the file will be deleted.