Stage 6/6: More features
Description

In this stage, you will parse fully-featured JSON documents. And after that, you will get the fully-featured XML/JSON converter.

There are some little details that you are missing: JSON has arrays, but every item of the array doesn't contain a key value; they just go one after another. To convert it to XML, you need to enclose these elements into the <element> tag. Also, you can detect an array in the XML if all the sub children (at least 2, a single child is not an array) have the same name (for instance, <transaction> in the example below). An empty JSON array should be converted to an empty string, just like a JSON empty object. An array inside an array does not have a key (which would be the opening and closing tag in XML) with which you can define this array. To avoid this, the array inside the array should be defined inside the <element> tags since every element in an array should be defined inside this tag.

Also, a comma doesn't have to follow the element, JSON lets you place it after any amount of spaces and new lines. You can see this in the example #3.

Another thing is that JSON can contain boolean values: true and false. During conversion, you should turn them into strings "true" and "false".

Also, XML supports values of attributes enclosed with single quotes, not only double quotes. For example, <xml attr="xml"></xml> and <xml attr='xml'></xml> mean the same and both of them are valid in XML.

At last, in this stage, you should skip any XML header that starts with <? and ends with ?>. There is no need for this in JSON.

Use examples as a reference as they cover all of the cases and tricky situations mentioned above.

Note that you do not need to output JSON or XML in a formatted form. You can print them in one line because the grader deletes all unnecessary whitespace characters before checking.

You should read from the file named test.txt. If you want to test your program, you should check it on some other file, because the contents of this file will be overwritten during testing and after testing the file will be deleted.